Fixed time changed logic
Whenever the time of a keyframe has changed, we need to check whether
the ordering remains the same. Previously, only the index of the key
that was changed was checked. But it could happen that the order of some
other keyframes have changed. Now, we check for the index of all
keyframes to see if the order has changed.